提币额度查询接口、额度初始化、临时额度消费/收入

+++ src/servlets/asset/withdrawal_quota.go
@@ -20,9 +20,9 @@ type withdrawQuotaParams struct {
 //}
 
 type withdrawQuotaResponse struct {
-	Day int `json:"day"`
-	Month int `json:"month"`
-	Casual int `json:"casual"`
+	Day int64 `json:"day"`
+	Month int64 `json:"month"`
+	Casual int64 `json:"casual"`
 }
